// types/interfaces.ts
interface UserState {
  email: string;
  token: string;
  role: string;
  isVerified: boolean;
}